Update docker build script
actually like check for failures.. before blowing away the old
containers.

+++ idsBuild.sh
@@ -27,8 +27,24 @@ echo Building the docker image...
 sed -i s/PLACEHOLDER_DOWNLOAD_PASSWORD/$DOWNLOAD_PASSWORD/g ./startup.sh
 sed -i s/PLACEHOLDER_DOWNLOAD_PASSWORD/$DOWNLOAD_PASSWORD/g ./Dockerfile
 ./docker/docker build -t gameon-logstash .
-./docker/docker stop -t 0 gameon-logstash
-./docker/docker rm gameon-logstash
-./docker/docker run -d -p $SL_HOST_IP:5043:5043 -p $SL_HOST_IP:514:514/udp -e ETCDCTL_ENDPOINT="http://etcd:4001" --link=etcd --name gameon-logstash gameon-logstash
-
-rm -rf dockercfg
+if [ $? != 0 ]
+  then
+    echo Docker Build failed.
+    curl -X 'POST' --silent --data-binary '{"text":"Docker Build for the logstash service has failed."}' $WEBHOOK > /dev/null
+    exit -2
+  else
+    echo Attempting to remove old containers.
+    ./docker/docker stop -t 0 gameon-logstash || true
+    ./docker/docker rm gameon-logstash || true
+    echo Starting new container.
+    ./docker/docker run -d -p $SL_HOST_IP:5043:5043 -p $SL_HOST_IP:514:514/udp -e ETCDCTL_ENDPOINT="http://etcd:4001" --link=etcd --name gameon-logstash gameon-logstash
+    if [ $? != 0 ]
+    then
+      echo Docker Run failed.
+      curl -X 'POST' --silent --data-binary '{"text":"Docker Run for the logstash service has failed."}' $WEBHOOK > /dev/null
+      exit -3
+    else
+      cd ..
+      rm -rf dockercfg
+    fi
+  fi
